data_management/datasync/requirements.txt
# DataSync Dependencies
# Database
psycopg2-binary>=2.9.0
SQLAlchemy>=1.4.0
alembic>=1.8.0

# Configuration
PyYAML>=6.0
python-dotenv>=0.19.0

# Logging
structlog>=22.0.0
colorlog>=6.0.0

# Utilities
click>=8.0.0
croniter>=1.3.0
requests>=2.28.0

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0